// firebase/functions/src/types/interface/geoDB/Location.ts
export interface Location   {
    locId: string;
    ags: string;
    ascii: string;
    name: string;
    lat: string;
    lon: string;
    amt: string;
    plz: string;
    telephonePrefix: string;
    inhabitants: string;
    area: string;
    kz: string;
    typ: string;
    level: string;
    of: string;
    invalid: string;
    country: string;
    [key: string]: string;
}